Delay snapshot creation a little bit
When drf-tracking is in use, the view is handled in a transaction. This
means the Snapshot isn't committed until after the view is done (and it
has been logged). Sometimes, the celery task runs before, and then it
fails to find the Snapshot.

Delay the task a tiny bit to overcome this. It's a bit racy, but it's
the best we have right now.